"평가 DB삭제 로직 구현"

+++ src/main/java/com/project/sundo_project/entity/Evaluation.java
@@ -18,7 +18,7 @@ public class Evaluation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "EvaluationID")
-    private int evaluationID;
+    private Long evaluationID;
 
     @Column(name = "ProjectId")
     private int projectID;
